Started a Makefile to run francinette tests

+++ include/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: fhongu <fhongu@student.42madrid.com>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/03 15:18:56 by fhongu            #+#    #+#             */
-/*   Updated: 2023/06/06 20:53:01 by fhongu           ###   ########.fr       */
+/*   Updated: 2023/06/06 23:10:10 by fhongu           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 # define FT_PRINTF_H
 # include <stdarg.h>
 # include <unistd.h>
-# include "libft/libft.h"
+# include "../libft/libft.h"
 
 typedef struct s_bflags
 {
@@ -34,6 +34,6 @@ typedef struct s_bflags
 int		ft_printf(const char *str, ...);
 void	printchar(int *counter, t_bflags bflags, int ch);
 void	printstr(int *counter, t_bflags bflags, char *str);
-void	printnum(int *counter, t_bflags bflags, long num);
+void	printnum(int *counter, t_bflags bflags, int num);
 
 #endif
